Hopeful final bugfix
Not the most elegant possible solution, and may have problems if multiple users connect at the exact same time, but it should work ok for now. Hopefully thats the last bugfix this branch will need. From here on out all the commits should relate to testing and the branch should be ready to be merged soon.